Perform renames in TASInput
