+++ src/Plugins.Win32.TASInput/Combo.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
-#include "Common.h"
-#include <Combo.h>
+#include "Common.hpp"
+#include <Combo.hpp>
 
 bool t_combo::uses_joystick() const
 {
