+++ src/Plugins.Win32.TASInput/ConfigDialog.hpp
