+++ src/Plugins.Win32.TASInput/GamepadManager.hpp
