+++ src/Plugins.Win32.TASInput/Helpers.hpp
