+++ src/Plugins.Win32.TASInput/JoystickControl.hpp
