+++ src/Plugins.Win32.TASInput/NewConfig.cpp
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
-#include "Common.h"
-#include <NewConfig.h>
+#include "Common.hpp"
+#include <NewConfig.hpp>
 #include <MiscHelpers.h>
-#include <Main.h>
+#include <Main.hpp>
 
 #define REG_SUBKEY L"Software\\N64 Emulation\\DLL\\TASDI"
 #define REG_CONFIG_VALUE L"Config"
