+++ src/Plugins.Win32.TASInput/NewConfig.hpp
